+++ packages/@leight/cursor-client/src/cursor/CursorControl.tsx
@@ -1,0 +1,16 @@
+import {
+    IQuerySchema,
+    IUseCursorCountQuery
+}                          from "@leight/query";
+import {PropsWithChildren} from "react";
+import {CursorProvider}    from "../context";
+
+export type ICursorControlProps<TQuerySchema extends IQuerySchema> = PropsWithChildren<{
+    useCountQuery: IUseCursorCountQuery<TQuerySchema>;
+}>;
+
+export const CursorControl = <TQuerySchema extends IQuerySchema>({children}: ICursorControlProps<TQuerySchema>) => {
+    return <CursorProvider>
+        {children}
+    </CursorProvider>;
+};